requirements.txt
python-dotenv>=0.21.0
google-generativeai>=0.3.0
faiss-cpu==1.12.0